+++ src/main/java/br/com/zalf/prolog/webservice/frota/veiculo/VeiculoDaoImpl.java
@@ -17,26 +17,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.*;
 
-public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
-
-    private static final String VEICULOS_BY_PLACA = "SELECT " +
-            "V.*, " +
-            "R.CODIGO AS COD_REGIONAL_ALOCADO, " +
-            "MV.NOME AS MODELO, " +
-            "EV.NOME AS EIXOS, " +
-            "EV.DIANTEIRO, " +
-            "EV.TRASEIRO, " +
-            "EV.CODIGO AS COD_EIXOS, " +
-            "tv.nome AS TIPO, " +
-            "MAV.NOME AS MARCA, " +
-            "MAV.CODIGO AS COD_MARCA  " +
-            "FROM VEICULO V JOIN MODELO_VEICULO MV ON MV.CODIGO = V.COD_MODELO " +
-            "JOIN EIXOS_VEICULO EV ON EV.CODIGO = V.COD_EIXOS " +
-            "JOIN VEICULO_TIPO TV ON TV.CODIGO = V.COD_TIPO AND TV.COD_UNIDADE = V.COD_UNIDADE " +
-            "JOIN MARCA_VEICULO MAV ON MAV.CODIGO = MV.COD_MARCA " +
-            "JOIN UNIDADE U ON U.CODIGO = V.COD_UNIDADE " +
-            "JOIN REGIONAL R ON U.COD_REGIONAL = R.CODIGO " +
-            "WHERE V.PLACA = ?;";
+public final class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
 
     public VeiculoDaoImpl() {
 
@@ -63,7 +44,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 throw new SQLException("Erro ao inserir o veículo");
             }
         } finally {
-            closeConnection(conn, stmt, null);
+            close(conn, stmt);
         }
         return true;
     }
@@ -86,7 +67,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 throw new SQLException("Erro ao atualizar o veículo");
             }
         } finally {
-            closeConnection(conn, stmt, null);
+            close(conn, stmt);
         }
         return true;
     }
@@ -105,7 +86,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 throw new SQLException("Erro ao atualizar o status do veículo com placa: " + placa);
             }
         } finally {
-            closeConnection(conn, stmt, null);
+            close(conn, stmt);
         }
     }
 
@@ -124,7 +105,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 throw new SQLException("Erro ao deletar o veículo");
             }
         } finally {
-            closeConnection(conn, stmt, null);
+            close(conn, stmt);
         }
         return true;
     }
@@ -174,7 +155,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 veiculos.add(createVeiculo(rSet));
             }
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return veiculos;
     }
@@ -195,7 +176,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                     "EV.DIANTEIRO, " +
                     "EV.TRASEIRO, " +
                     "EV.CODIGO AS COD_EIXOS, " +
-                    "tv.nome AS TIPO, " +
+                    "TV.NOME AS TIPO, " +
                     "MAV.NOME AS MARCA, " +
                     "MAV.CODIGO AS COD_MARCA  "
                     + "FROM VEICULO V JOIN MODELO_VEICULO MV ON MV.CODIGO = V.COD_MODELO "
@@ -213,7 +194,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 veiculos.add(veiculo);
             }
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return veiculos;
     }
@@ -234,7 +215,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
             conn = getConnection();
             return internalGetVeiculoByPlaca(conn, placa, withPneus);
         } finally {
-            closeConnection(conn);
+            close(conn);
         }
     }
 
@@ -245,7 +226,24 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
         PreparedStatement stmt = null;
         ResultSet rSet = null;
         try {
-            stmt = conn.prepareStatement(VEICULOS_BY_PLACA);
+            stmt = conn.prepareStatement("SELECT " +
+                    "V.*, " +
+                    "R.CODIGO AS COD_REGIONAL_ALOCADO, " +
+                    "MV.NOME AS MODELO, " +
+                    "EV.NOME AS EIXOS, " +
+                    "EV.DIANTEIRO, " +
+                    "EV.TRASEIRO, " +
+                    "EV.CODIGO AS COD_EIXOS, " +
+                    "tv.nome AS TIPO, " +
+                    "MAV.NOME AS MARCA, " +
+                    "MAV.CODIGO AS COD_MARCA  " +
+                    "FROM VEICULO V JOIN MODELO_VEICULO MV ON MV.CODIGO = V.COD_MODELO " +
+                    "JOIN EIXOS_VEICULO EV ON EV.CODIGO = V.COD_EIXOS " +
+                    "JOIN VEICULO_TIPO TV ON TV.CODIGO = V.COD_TIPO " +
+                    "JOIN MARCA_VEICULO MAV ON MAV.CODIGO = MV.COD_MARCA " +
+                    "JOIN UNIDADE U ON U.CODIGO = V.COD_UNIDADE " +
+                    "JOIN REGIONAL R ON U.COD_REGIONAL = R.CODIGO " +
+                    "WHERE V.PLACA = ?;");
             stmt.setString(1, placa);
             rSet = stmt.executeQuery();
             if (rSet.next()) {
@@ -259,100 +257,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 throw new IllegalStateException("Erro ao buscar veículo com a placa: " + placa);
             }
         } finally {
-            closeConnection(null, stmt, rSet);
-        }
-    }
-
-    @Override
-    public List<TipoVeiculo> getTipoVeiculosByUnidade(Long codUnidade) throws SQLException {
-        List<TipoVeiculo> listTipo = new ArrayList<>();
-        Connection conn = null;
-        PreparedStatement stmt = null;
-        ResultSet rSet = null;
-        try {
-            conn = getConnection();
-            stmt = conn.prepareStatement("SELECT * FROM VEICULO_TIPO WHERE COD_UNIDADE = ? AND STATUS_ATIVO = TRUE");
-            stmt.setLong(1, codUnidade);
-            rSet = stmt.executeQuery();
-            while (rSet.next()) {
-                listTipo.add(new TipoVeiculo(rSet.getLong("CODIGO"), rSet.getString("NOME")));
-            }
-        } finally {
-            closeConnection(conn, stmt, rSet);
-        }
-        return listTipo;
-    }
-
-    @Override
-    public boolean insertTipoVeiculo(TipoVeiculo tipoVeiculo, Long codUnidade) throws SQLException {
-        Connection conn = null;
-        PreparedStatement stmt = null;
-        try {
-            conn = getConnection();
-            stmt = conn.prepareStatement("INSERT INTO VEICULO_TIPO(COD_UNIDADE, NOME, STATUS_ATIVO) VALUES (?,?,?)");
-            stmt.setLong(1, codUnidade);
-            stmt.setString(2, tipoVeiculo.getNome());
-            stmt.setBoolean(3, true);
-            int count = stmt.executeUpdate();
-            if (count == 0) {
-                throw new SQLException("Erro ao cadastrar o tipo de veículo");
-            }
-        } finally {
-            closeConnection(conn, stmt, null);
-        }
-        return true;
-    }
-
-    @Override
-    public TipoVeiculo getTipoVeiculo(Long codTipo, Long codUnidade) throws SQLException {
-        Connection conn = null;
-        PreparedStatement stmt = null;
-        ResultSet rSet = null;
-        try {
-            conn = getConnection();
-            stmt = conn.prepareStatement("SELECT * FROM veiculo_tipo WHERE codigo = ? AND cod_unidade = ?");
-            stmt.setLong(1, codTipo);
-            stmt.setLong(2, codUnidade);
-            rSet = stmt.executeQuery();
-            if (rSet.next()) {
-                TipoVeiculo tipo = new TipoVeiculo();
-                tipo.setCodigo(rSet.getLong("CODIGO"));
-                tipo.setNome(rSet.getString("NOME"));
-                return tipo;
-            }
-        } finally {
-            closeConnection(conn, stmt, rSet);
-        }
-        return null;
-    }
-
-    @Override
-    public boolean updateTipoVeiculo(TipoVeiculo tipo, Long codUnidade) throws SQLException {
-        Connection conn = null;
-        PreparedStatement stmt = null;
-        try {
-            conn = getConnection();
-            stmt = conn.prepareStatement("UPDATE veiculo_tipo SET nome = ? WHERE codigo = ? AND cod_unidade = ?;");
-            stmt.setString(1, tipo.getNome());
-            stmt.setLong(2, tipo.getCodigo());
-            stmt.setLong(3, codUnidade);
-            return stmt.executeUpdate() > 0;
-        } finally {
-            closeConnection(conn, stmt, null);
-        }
-    }
-
-    public boolean deleteTipoVeiculo(Long codTipo, Long codUnidade) throws SQLException {
-        Connection conn = null;
-        PreparedStatement stmt = null;
-        try {
-            conn = getConnection();
-            stmt = conn.prepareStatement("DELETE FROM veiculo_tipo WHERE codigo = ? AND cod_unidade = ?");
-            stmt.setLong(1, codTipo);
-            stmt.setLong(2, codUnidade);
-            return stmt.executeUpdate() > 0;
-        } finally {
-            closeConnection(conn, stmt, null);
+            close(stmt, rSet);
         }
     }
 
@@ -375,7 +280,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 eixos.add(eixo);
             }
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return eixos;
     }
@@ -398,6 +303,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
         }
     }
 
+    @Deprecated
     @Override
     public List<Marca> getMarcaModeloVeiculoByCodEmpresa(Long codEmpresa) throws SQLException {
         Connection conn = null;
@@ -447,7 +353,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
             marca.setModelos(modelos);
             marcas.add(marca);
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return marcas;
     }
@@ -566,7 +472,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 return modelo;
             }
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return null;
     }
@@ -585,7 +491,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
             stmt.setLong(4, codUnidade);
             return stmt.executeUpdate() > 0;
         } finally {
-            closeConnection(conn, stmt, null);
+            close(conn, stmt);
         }
     }
 
@@ -601,7 +507,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
             stmt.setLong(2, codUnidade);
             return stmt.executeUpdate() > 0;
         } finally {
-            closeConnection(conn, stmt, null);
+            close(conn, stmt);
         }
     }
 
@@ -618,7 +524,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 total = rSet.getInt("COUNT");
             }
         } finally {
-            closeConnection(null, stmt, rSet);
+            close(stmt, rSet);
         }
         return total;
     }
@@ -631,10 +537,8 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
         List<String> placas = new ArrayList<>();
         try {
             conn = getConnection();
-            stmt = conn.prepareStatement("SELECT V.PLACA FROM veiculo V JOIN veiculo_tipo VT ON V.cod_unidade = VT.cod_unidade\n" +
-                    "AND V.cod_tipo = VT.codigo\n" +
-                    "WHERE VT.cod_unidade = ? AND VT.codigo::TEXT LIKE ? " +
-                    "ORDER BY V.PLACA");
+            stmt = conn.prepareStatement("SELECT V.PLACA FROM VEICULO V JOIN VEICULO_TIPO VT ON V.COD_TIPO = VT.CODIGO " +
+                    "WHERE V.COD_UNIDADE = ? AND VT.CODIGO::TEXT LIKE ? ORDER BY V.PLACA;");
             stmt.setLong(1, codUnidade);
             stmt.setString(2, codTipo);
             rSet = stmt.executeQuery();
@@ -642,7 +546,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 placas.add(rSet.getString("placa"));
             }
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return placas;
     }
@@ -654,7 +558,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
             conn = getConnection();
             return internalGetDiagramaVeiculoByPlaca(conn, placa);
         } finally {
-            closeConnection(conn);
+            close(conn);
         }
     }
 
@@ -664,28 +568,6 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
         return internalGetDiagramaVeiculoByPlaca(conn, placa);
     }
 
-    @NotNull
-    private Optional<DiagramaVeiculo> internalGetDiagramaVeiculoByPlaca(@NotNull final Connection conn,
-                                                                        @NotNull final String placa) throws SQLException {
-        PreparedStatement stmt = null;
-        ResultSet rSet = null;
-        try {
-            stmt = conn.prepareStatement("SELECT vd.*\n" +
-                    "FROM veiculo v JOIN veiculo_tipo vt on v.cod_tipo = vt.codigo and v.cod_unidade = vt.cod_unidade\n" +
-                    "JOIN veiculo_diagrama vd on vd.codigo = vt.cod_diagrama\n" +
-                    "WHERE v.placa = ?");
-            stmt.setString(1, placa);
-            rSet = stmt.executeQuery();
-            if (rSet.next()) {
-                return createDiagramaVeiculo(rSet, conn);
-            }
-        } finally {
-            closeStatement(stmt);
-            closeResultSet(rSet);
-        }
-        return Optional.empty();
-    }
-
     @Override
     public Optional<DiagramaVeiculo> getDiagramaVeiculoByCod(@NotNull Short codDiagrama) throws SQLException {
         Connection conn = null;
@@ -693,10 +575,10 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
         ResultSet rSet = null;
         try {
             conn = getConnection();
-            stmt = conn.prepareStatement("SELECT *\n" +
-                    "FROM veiculo_diagrama AS vd\n" +
-                    "  JOIN veiculo_tipo AS vt\n" +
-                    "    ON vd.codigo = vt.cod_diagrama\n" +
+            stmt = conn.prepareStatement("SELECT * " +
+                    "FROM veiculo_diagrama AS vd " +
+                    "  JOIN veiculo_tipo AS vt " +
+                    "    ON vd.codigo = vt.cod_diagrama " +
                     "WHERE vd.codigo = ?");
             stmt.setShort(1, codDiagrama);
             rSet = stmt.executeQuery();
@@ -704,7 +586,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 return createDiagramaVeiculo(rSet, conn);
             }
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return Optional.empty();
     }
@@ -723,7 +605,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 createDiagramaVeiculo(rSet, conn).ifPresent(diagramas::add);
             }
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return diagramas;
     }
@@ -746,7 +628,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 throw new SQLException("Erro ao aplicar o pneu " + codPneu + " ao veículo " + placa);
             }
         } finally {
-            closeStatement(stmt);
+            close(stmt);
         }
     }
 
@@ -767,10 +649,32 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 throw new SQLException("Erro ao remover o pneu " + codPneu + " da placa " + placa);
             }
         } finally {
-            closeStatement(stmt);
+            close(stmt);
         }
     }
 
+    @NotNull
+    private Optional<DiagramaVeiculo> internalGetDiagramaVeiculoByPlaca(@NotNull final Connection conn,
+                                                                        @NotNull final String placa) throws SQLException {
+        PreparedStatement stmt = null;
+        ResultSet rSet = null;
+        try {
+            stmt = conn.prepareStatement("SELECT VD.* " +
+                    "FROM VEICULO V JOIN VEICULO_TIPO VT ON V.COD_TIPO = VT.CODIGO " +
+                    "JOIN VEICULO_DIAGRAMA VD ON VD.CODIGO = VT.COD_DIAGRAMA " +
+                    "WHERE V.PLACA = ?");
+            stmt.setString(1, placa);
+            rSet = stmt.executeQuery();
+            if (rSet.next()) {
+                return createDiagramaVeiculo(rSet, conn);
+            }
+        } finally {
+            close(stmt, rSet);
+        }
+        return Optional.empty();
+    }
+
+    @NotNull
     private Optional<DiagramaVeiculo> createDiagramaVeiculo(ResultSet rSet, Connection conn) throws SQLException {
         return Optional.of(new DiagramaVeiculo(
                 rSet.getShort("CODIGO"),
@@ -779,6 +683,7 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 rSet.getString("URL_IMAGEM")));
     }
 
+    @NotNull
     private Set<EixoVeiculo> getEixosDiagrama(int codDiagrama, Connection conn) throws SQLException {
         PreparedStatement stmt = null;
         ResultSet rSet = null;
@@ -799,11 +704,12 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
                 eixos.add(eixoVeiculo);
             }
         } finally {
-            closeConnection(null, stmt, rSet);
+            close(stmt, rSet);
         }
         return eixos;
     }
 
+    @NotNull
     private Veiculo createVeiculo(ResultSet rSet) throws SQLException {
         final Veiculo veiculo = new Veiculo();
         veiculo.setPlaca(rSet.getString("PLACA"));
@@ -843,11 +749,11 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
         return veiculo;
     }
 
+    @NotNull
     public List<Veiculo> getVeiculoKm(Long codUnidade, String placa, String codTipo) throws SQLException {
         Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
-        List<Veiculo> veiculos = new ArrayList<>();
         Veiculo v = null;
         try {
             conn = getConnection();
@@ -857,15 +763,16 @@ public class VeiculoDaoImpl extends DatabaseConnection implements VeiculoDao {
             stmt.setString(2, codTipo);
             stmt.setString(3, placa);
             rSet = stmt.executeQuery();
+            final List<Veiculo> veiculos = new ArrayList<>();
             while (rSet.next()) {
                 v = new Veiculo();
                 v.setPlaca(rSet.getString("placa"));
                 v.setKmAtual(rSet.getLong("km"));
                 veiculos.add(v);
             }
+            return veiculos;
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
-        return veiculos;
     }
 }