+++ src/main/java/br/com/zalf/prolog/webservice/integracao/operacoes/OperacoesIntegradasVeiculo.java
@@ -15,7 +15,7 @@ interface OperacoesIntegradasVeiculo {
     List<Veiculo> getVeiculosAtivosByUnidade(@NotNull final Long codUnidade, @Nullable final Boolean ativos) throws Exception;
 
     @NotNull
-    List<TipoVeiculo> getTiposVeiculosByUnidade(@NotNull final Long codUnidade) throws Exception;
+    List<TipoVeiculo> getTiposVeiculosByEmpresa(@NotNull final Long codEmpresa) throws Throwable;
 
     @NotNull
     List<String> getPlacasVeiculosByTipo(@NotNull final Long codUnidade, @NotNull final String codTipo) throws Exception;